Fixed some compatibility issues related to Py2 and Py3

+++ nurbs/utilities.py
@@ -88,7 +88,7 @@ def find_span(degree=0, knotvector=(), knot=0):
             high = mid
         else:
             low = mid
-        mid = (low + high) / 2
+        mid = int((low + high) / 2)
 
     return mid
